feat: 횟수 타입 Integer 대신 Long 사용

+++ src/main/java/racingcar/validation/RoundValidator.java
@@ -22,7 +22,7 @@ public class RoundValidator implements Validator {
     }
 
     private void validateRange(String value) {
-        if (0 > Integer.parseInt(value)) {
+        if (0 > Long.parseLong(value)) {
             throw new IllegalArgumentException();
         }
     }
